requirements.txt
opencv-python==4.10.0.84
pillow==10.4.0
python-doctr==0.8.1
numpy==1.26.4
PyMuPDF==1.24.7
fastapi==0.111.0
slowapi==0.1.9
nanoid==2.0.0
